feat: reorganize dashboard types to avoid circular dependencies and improve structure
