+++ packages/mosaic/src/chart-types/base-types.ts
@@ -12,13 +12,10 @@ import {ChartConfig, ChartType} from './chart-config';
 import {RetainedVgPlotChart} from '../VgPlotChart';
 import type {Selection} from '@uwdata/mosaic-core';
 import type {
-  MosaicDashboardEntry,
   ProfilerPanelConfig,
   TextPanelConfig,
-} from '../dashboard/dashboard-types';
-
-// Re-export ChartType for convenience
-export type {ChartType};
+} from '../dashboard/core-types';
+import type {MosaicDashboardEntry} from '../dashboard/dashboard-types';
 
 /**
  * Column info passed to chart builder UI
