+++ packages/mosaic/src/dashboard/defaultPanelActions.ts
@@ -1,4 +1,4 @@
-import {type MosaicDashboardAddPanelAction} from './dashboard-types';
+import {type MosaicDashboardAddPanelAction} from './action-types';
 import {addTextPanelAction} from '../text/addTextPanelAction';
 import {addProfilerPanelAction} from '../profiler/addProfilerPanelAction';
 import {addChartPanelAction} from '../chart/addChartPanelAction';
