+++ packages/mosaic/src/profiler/addProfilerPanelAction.ts
@@ -1,5 +1,5 @@
 import {TableProperties} from 'lucide-react';
-import {type MosaicDashboardAddPanelAction} from '../dashboard/dashboard-types';
+import {type MosaicDashboardAddPanelAction} from '../dashboard/action-types';
 import {
   createMosaicDashboardProfilerPanelConfig,
   MOSAIC_DASHBOARD_PROFILER_PANEL_TYPE,
